added config check

+++ multinode.sh
@@ -211,6 +211,12 @@ fi
 function upnp_disable() {
 
  echo -e ""
+ if [[ ! -f /home/$USER/zelflux/config/userconfig.js ]]; then
+       echo -e "${WORNING} ${CYAN}Missing FluxOS configuration file - install/re-install Flux Node...${NC}" 
+       echo -e ""
+       exit
+ fi
+ 
  if [[ -f /home/$USER/.fluxbenchmark/fluxbench.conf ]]; then
  echo -e "${ARROW} ${CYAN}Removing FluxOS UPnP configuration.....${NC}"
  sudo rm -rf /home/$USER/.fluxbenchmark/fluxbench.conf
